pyproject.toml
[tool.poetry]
name = "invdb-py"
version = "0.1.0"
description = ""
authors = ["Dillon Brown <dillon.j.brown@saic.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
psycopg2-binary = "^2.9.9"
openpyxl = "^3.1.2"
multithreading = "^0.2.0"
loguru = "^0.7.2"
requests = "^2.31.0"
chalice = "^1.29.0"
flask = "^3.0.0"
waitress = "^2.1.2"
celery = "^5.3.6"
fuzzywuzzy = "^0.18.0"
python-levenshtein = "^0.25.0"
pytest = "^8.3.2"
boto3 = "^1.35.66"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
debugpy = "^1.8.9"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
